feat(steppers): Implement IStringStepper; Create Mutable Implementation; Add Numerous Helpful Extension Methods
This commit adds `StringStepper`, `MutableStringStepper` and corresponding interface classes providing the ability to traverse and manipulate the content of a string. Notably, the implementation includes functions to Jump to a specific position, reset the stepper and dispose the object. Furthermore, additional extensions methods and options classes have been added to extend this functionality and provide a more granular level of control over string manipulation operations. These changes are necessary to allow more dynamic and flexible handling of strings within the LangAssembler project.