feat(lexer): Finish Up TokenSetLexer
The Lexer classes have been refactored and moved to a 'Base' subdirectory. This measure was necessary to improve project organization. A new method 'GetNextMatch' has been introduced to 'BaseLexer' and 'TokenSetLexer' to identify the next token type within a given input buffer. This enhances the token matching logic and makes it more flexible for subclasses of BaseLexer. The new class 'TokenSetExtensions' has been created but not finished to aid in fetching the token sets associated with a given Lexer.