feat(parser): Create Parser Project
In this commit, two new interfaces `IParserContext` and `IPaser` were added. These interfaces provide a contract for the parsing context and a parser respectively. The parser transforms a series of tokens into an Abstract Syntax Tree, using the provided context. This addition is integral to parsing functionality.

Additionally, a `Parser` base class and `ParserContextExtensions` were created which included functionality required for parsing like `ShouldContinue` and `MarkEnd`.

The file `TokenSetLexer` was updated to reflect the change in class name from `BaseLexer` to `Lexer`. The existing `BaseLexer` was renamed to `Lexer`, and all the associated dependencies were also updated.

Also, a new project `LangAssembler.Parser` was added to the solution and referenced in the `LangAssembler.Lexer` project. This enables the lexical analysis processes to make use of the new parsing components.