feat(processor): Unicode Support
This commit primarily updates the StringProcessor and Document classes to better manage content. It changes the CurrentChar and PreviousChar from char? to int? to provide better compatibility with Unicode characters. The access modifiers of several properties in the Document class have been changed to private and certain properties were renamed for better readability. Additional methods were also added for document line management and positioning of the cursor. BinaryReader in DocumentReader was replaced with StreamReader for better string handling. Minor changes were also made to methods in IStringProcessor, DocumentProcessor, Lexer and DocumentLineInfo. In addition to these changes, a new interface IDocumentLineInfo has been added to represent line information in a document. These changes improves overall data handling and prepares the ground for forthcoming features.