feat(document): Add more functionality to DocumentProcessor and improve code documentation
In StringProcessor.cs, an 'override ToString()' method was added for returning processed content.

The majority of changes were made in DocumentProcessor.cs. The StreamReader is replaced with a BinaryReader for document reading; Constructors were modified and added to accommodate a BinaryReader. Finalizer and few other methods are added to the class. The major change is the addition of method JumpTo() which allows the document position to be changed in processing with correct state settings.

These changes reflect a necessity of handling different types of document data in the language assembler project.

IDocumentProcessor has been adjusted accordingly to include the JumpTo method.

Minor cleanup was done in document extensions and specific method documentation / comments were improved in DocumentLineInfo.cs.

The aim behind these changes is to improve the usability, flexibility, and comprehension of the code.