feat(generator/lexer): Finish Prototype Lexer Grammar; Create Models To Map Language To
Added Model folder to Lexer with relevant Lexer models for metadata, literals, and elements in mind. This was necessary to represent abstract syntax tree for lexer grammar. Additionally, unused Class1.cs file was removed, while the grammar was refined with the introduction of metadata and literals, which are necessary to represent any Lex script. Further divided literals into base types of String, Hex, and RuleName to allow a varied input for users.