feat(lexer): Refactor token matching mechanism for lexer
This major refactor changes the way tokens are matched in the lexer. It introduces a new TokenMatcherDelegate and modifies the ITokenType interface to work with this new delegate mechanism, allowing token match implementations to be passed directly into token types.

This approach removes the need for each token type to implement the match function individually, reducing redundancy in code. All corresponding uses of the old matching mechanism were updated as well.

Additional changes include a minor change in Language.cs to simplify language registration and set UTF8 as the default encoding for ALexLanguage.